chore: inline detailed_formatter_helpers.go into detailed_formatter.go
The eight pure utility functions in detailed_formatter_helpers.go
(stripWhitespace, visualizeWhitespace, yamlTypeName, formatCount,
pluralize, formatTimestamp, formatCommaSeparated, formatDetailedValue,
isWhitespaceOnlyChange) were used only by the detailed formatter.
Moving them into detailed_formatter.go removes one file from the
package without losing the domain seams: detailed_formatter_render.go
(YAML rendering) and detailed_formatter_linediff.go (Myers algorithm)
remain separate, since those have real distinct concerns.
